+++ CHANGELOG.md
@@ -20,6 +20,8 @@
   * Support filtering group issues by their type.
   * Support filtering out external users when listing users.
   * Support setting and editing the `raw` flag on project CI variables.
+  * Support setting and editing the `description` field on project CI
+    variables.
 
 ## Changes